Move MMR shared tests to their correct file
