+++ twenty-first/src/util_types/mmr/mmr_membership_proof.rs
@@ -1,6 +1,5 @@
 use itertools::Itertools;
 use serde::{Deserialize, Serialize};
-
 use std::collections::hash_map::RandomState;
 use std::collections::hash_set::Intersection;
 use std::collections::{HashMap, HashSet};
